pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brickwell-health"
version = "0.1.0"
description = "Australian Private Health Insurance Simulator"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Brickwell Health", email = "dev@brickwellhealth.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "simpy>=4.1.1",
    "sqlalchemy>=2.0.25",
    "psycopg[binary]>=3.1.17",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "numpy>=1.26.3",
    "faker>=22.0.0",
    "holidays>=0.40",
    "structlog>=24.1.0",
    "tqdm>=4.66.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-xdist>=3.5",
    "factory-boy>=3.3",
    "mypy>=1.8",
    "ruff>=0.2",
    "pre-commit>=3.6",
    "types-PyYAML>=6.0.12",
    "types-tqdm>=4.66.0",
]

[project.scripts]
brickwell = "brickwell_health.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["brickwell_health*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
